requirements.txt
streamlit
transformers
sentence-transformers
faiss-cpu
PyMuPDF
pdfplumber
langchain
peft
trl
torch
accelerate
datasets
huggingface_hub
scikit-learn
transformers
sentence-transformers
torch
faiss-cpu
numpy
pypdf
langchain
datasets
accelerate
bitsandbytes
peft
trl
gradio
streamlit